requirements.txt
sqlalchemy==2.0.23
requests==2.31.0
slack-sdk==3.23.0
openai==1.3.0
python-dotenv==1.0.0
hypothesis==6.88.0
pytest==7.4.3
pytest-cov==4.1.0
strands-sdk==0.1.0
prometheus-client==0.19.0
